Officially remove the worst part of UnsafeFormulaManager.
Long ago, we had no way to get the type of a formula from a solver,
thus UnsafeFormulaManager.getArg() used to return untyped formulas
(i.e., plain Formula instances instead of instances of a subtype
like BooleanFormula). This was changed some time ago,
now getArg() returns correctly typed formulas, and all other method
of this interface should do so, too, but it was never officially
declared.

There was also a method typeFormula() that could be used to simply
set any type on any formula, which was used to work around get
limitation of getArg(). Now it is not needed anymore and thus removed.