moving the switch-case-code from the visitor into the BooleanFormulaMgr.
The code to switch between different kind of operations is now intered
in the BooleanFormulaMgr and is not visible for the user.

This is a first step towards removing the public API to get the type of
a formula, see #14.

Currently the code is very similar in all solver-specific classes.