Update cvc4 ant target to work without git, but with downloadable releases.
For this to work the source tarball has to be downloaded and compiled.
Additionally to these changes, we run strip on the built libraries so
their size will be much smaller in the future. As the libcvc4jni.so is
currently linked against another file in a certain path this is quite
useless, so until I have a fix for this I removed the .so files from
the ivy repository. The jar is necessary to have no compilation errors
inside java-smt.