Change to FormulaVisitor Quantifier API
When de-Bruijn indexes are used (and they are used all the time),
quantifier does not have to know which variables it bounds --- it can
be derived from the de-Bruijn indexes of bound variables.

In case of Princess the quantifier node indeed does not know,
and the only way to get bound variables is to walk the AST manually.

Moreover, the previous API had issues:

- from visitBoundVariable it was impossible to figure out which
quantiifer it belongs to.

- List of bound vars returned to the client wasn't of "bound" type,
   but the returned variables were constants instead.

Now instead of providing a list of bound variables of unclear type, we
provide a constructor to directly replace the quantifier body instead.