Fix array type detection in SMTInterpol
SMTInterpol produces terms of the form `(as @0 (Array Int Int))` that
have not been detected as array types yet. This caused
`IllegalArgumentException`s when we try to create the
`SmtInterpolModel`.

We now detect this type of terms as array types, but the selection of
the array values for the model are still open (as with any other
solver).