Removes methods for formula introspection.
As a result, UnsafeFormulaManager only contains "visit" implementation,
which exists in a separate module purely for convenience in code
sharing between different managers for the same solver.
Thus it is no longer exposed and renamed to the IntrospectionManager
(however, the interface is removed and only the abstract base class
remains).

Closes #14.