Update IntelliJ configuration files
Introduces major changes in the IDE configuration files! Let me know, if
you encounter any problems with this configuration.

We add a configuration for the ANT plugin that tells the plugin that
`build.xml` is the default ANT file. We further set the JUnit version to
4, such that the test generator knows of this.

The major changes regard Checkstyle and Findbugs. You need to have the
IntelliJ plugins `Checkstyle-IDEA` and `Findbugs-IDEA` installed. This
configuration works with IntelliJ IDEA Ultimate 2016.1.1,
Checkstyle-IDEA 4.27.0, and Findbugs-IDEA 0.9.997.

The default checkstyle configuration provided in the project settings is
used as a default; for unit tests you have to select to appropriate
configuration in order to omit wrong warnings.