Added a unit-test for interpolation with empty groups
Corner cases of interpolation are that group A or B are empty. All
solvers besides SMTInterpol are able to work with that (just by removing
too strict constraints in Z3, MathSAT5 and Princess could do it by
default). For SMTInterpol I added some extra handling that just gives
back the correct answer: TRUE for empty interpolation group A and FALSE
for empty interpolation group B.