Use Z3 JNI bindings instead of our own ones.
This gives us the best of both worlds: optimal performance and not
having a technical debt of having to update the JNI generation script
each time Z3 API changes.
Their constant declarations are better as well, done as proper Java
enums.
A logical next step is to get rid of libz3j.so altogether, and of the
python script used to generate it.