Use the new SMTInterpol from Maven.
This version uses its own logging interface instead of log4j,
so we need a new logging adaptor.

This version breaks all interpolation because it rejects interpolation
queries when arrays are enabled in the theory (which we always do),
even if there are no arrays in the formula
(cf. ultimate-pa/smtinterpol#22).