Update compilation process for MathSAT5 (closes #75).
1. Bind libstdc++ statically into libmathsatj, because the MathSAT5
binary needs a symbol from GLIBCXX_3.4.20, and this is not available
on systems such as Ubuntu 14.04.
With this change, a recent libstdc++ is still necessary for compiling
the bindings, but not for using the resulting binary.
This adds about 1.5 MB of code to the binary.

2. Mark all symbols as hidden in the binary except the JNI symbols, which start with "Java_".
This shouldn't hurt and maybe it avoids problems on systems with
differing libstdc++ versions,
because otherwise we would export all of libstdc++'s symbols, too.
(This also hides the GMP symbols that result from statically binding
 libgmp, which is probably also a good idea.)
This shrinks the binary by about 1.5 MB, maybe because the linker can
now completely remove unused functions.
Everything seems to work, though.

Maybe the second point is a good idea for the Z3 bindings, too?