Copy .travis.yml from java-common-lib.
It works there nicely, provides all the features of the current one,
and is easier.

Compared to the current version, it does not add a useless builder that
just runs "true", and it splits unit tests and Javadoc.
The builder that just builds is also removed, because the same build is
also done by the builders for unit tests and FindBugs, so it is
redundant.
This also fixes a bug about deploying coverage twice and Javadoc never.