fix wrong example and its wrong documentation.
Z3 seems to have a bug for optimization,
such that the returned `handle` contains a wrong optimal value.
The model seems to be correct.
The bug is reported at https://github.com/Z3Prover/z3/issues/1326 .

Thus currently the output of the running example and the `ideal solution` do not match.