Remove memcpy wrapper for Z3
Even Ubuntu 12.04 supports glibc 2.14,
and it is one step towards being able to use the official Z3 binaries.

For now, we still keep the own linking of libz3.so from libz3.a
because otherwise libz3.so cannot be loaded from within Java.

Note that we can keep the memcpy wrapper for MathSAT a little bit longer
because for MathSAT we need to compile our own binaries anyway.