Add an option for approximating non-linear arithmetic with UFs.
Non-linear arithmetic is not supported by some solvers,
and even if supported it is not always desired to use it
due to incompleteness and complexity of the solver.
We now provide the possibility to approximate non-linear arithmetic
for integers and rationals with UFs,
either always (to avoid use of non-linear arithmetic)
or as fallback (to use non-linear arithmetic where supported by the solver).

The default is to not approximate anything,
i.e., throw exceptions if a non-linear operation is attempted
and the solver does not support it (same behavior as before).

This fixes #131.