introduce a new ProverOption for generating all satisfying assignments.
The option GENERATE_MODELS seems to cause overhead in some solvers,
i.e., Mathsat5 in CPAchecker (TerminationAnalysis).
As only Z3 uses model-generation for allSat-generation,
a separate option is better here.

This commit removes the need of GENERATE_MODELS for AllSat computation
and adds a new option GENERATE_ALL_SAT for this case.