Update some dependencies, including Guava and Truth
To do this we need to disable null checks of SpotBugs,
but there is no progress in SpotBugs on supporting the new annotations,
and we do not want to use the old dependencies infinitely.

Upgrading Truth triggers a refactoring of the Truth Subjects that we have,
and while at it I also wrote tests for these subjects.
Truth not only has nice utilities for writing tests,
it also has nice utilities for writing test utilities
and for testing these test utilities.