Declare all Formula and FormulaType classes as @Immutable
This will allow users of JavaSMT who also use Error-Prone
to use formula instances in places where they need an immutable type.

For AbstractFormula and FunctionDeclarationImpl we take a shortcut.
Instead of declaring AbstractFormula as
`@Immutable class AbstractFormula<@ImmutableTypeParameter TFormulaInfo>`
(which would require `TFormulaInfo` to be an immutable type),
we declare it as
`@Immutable(containerOf = "TFormulaInfo") class AbstractFormula<TFormulaInfo>`
(which does not restrict `TFormulaInfo`).

The former would be more strict (and actually be the desired semantics),
but would mean that we also need to declare `TFormulaInfo` in
FormulaCreator as `@ImmutableTypeParameter`, which Error-Prone lets us
do only if FormulaCreator is `@Immutable`, which is not what we want.