Filter irrelevant operands from n-ary and/or for Z3
An extended form of this was was rejected in the discussion of
https://github.com/sosy-lab/java-smt/commit/36e47b5beb8e67938bf627367fd854a97a8e43f3
and reverted in 2a10a87e.
But even before that, the same simplification had been added to
BooleanFormulaTransformationVisitor in 5fe5d916, which is clearly worse:
- If the optimization is relevant in the visitor,
  then why shouldn't it be always relevant?
  We also do it for binary and/or, why not for n-ary?
- Doing it in the visitor penalizes solvers that do the optimization
  themselves anyway.
- Even for Z3, the optimization in the visitor needs an extra copy,
  whereas the optimization in Z3BooleanFormulaManager can do it with
  basically no extra cost. We do it without any additional allocation
  and just inline Longs.toArray with some extra conditions.

So now we can revert 5fe5d916 and add more tests for simplification.