Simplify n-ary and/or with duplicate operands like and(x, y, x)
SMTInterpol already does this, now we do it for all other solvers as well.
This basically re-adds 36e47b5b (which was only for Z3),
but in a more memory-efficient way (Stream.distinct() typically uses
a regular HashSet, the old commit used a LinkedHashSet).

The performance and memory consumption are O(n) (amortized).

Cf. #73