Attempt at better bv function handling
Need to find a way to handle yices_bitextract/ bit(t i).

SolverVisitorTest/bvvisit fails because it tries to get the args of a
bit extract with term_child. But bit extract is a projection and needs
proj_arg/proj_index to get the args.

SolverVisitorTest/bitvectorIdVisit fails because the multiplication is
an array of (and(bit t i) (bit ti)), which currently has
FunctionKind.OTHER.