Workaround for missing AND_Constructor/ Let JavaSMT handle substitution
Yices has no explicit AND constructor as it transforms AND(x,...) into
NOT(OR(NOT(X),NOT(...)).

This led to calling NOT with more than one parameter in
callFunctionImpl.

To workaround this problem the constructor YICES_AND = Integer.MAX_VALUE
was added. This values was chosen to ensure no collisions with actual
constructors occur.

Let JavaSMT handle the substitution as yices_subst_term() can only
substitute variables/uninterpreted terms.