add JUnit tests for Boolector dump.
The first test dumps a plain boolean symbol in invalid SMTLIB format.
The assertion is missing (see out-commented line in the JUnit test).

The second test fail sometimes because of a non-deterministic char at the end of the String.
The char seems to be "\0" and appears sometimes, even if we "strip" the string in the
BoolectorFormulaManager.
Additionally, SMTLIB is invalid, because the declarations of symbols is missing
and the formula is just dumped without any context like "(assert FORMULA)".