split jar file into the core JavaSMT files and the component for Yices2.
(initial try)

As Yices2 is under the GPLv3 license,
we cannot ship the binary directly,
but need to extract it into a seperate jar file.
Currently, the sources and JavaDoc are kept together.

This change needs further work, and we might want to have
a consistent naming scheme and splitted sources and JavaDoc,