Improve Truth assertion in test
The power of Truth comes from using its assertions properly.
Code like "assertThat(foo.equals(bar)).isTrue()" and
"assertThat(foo.compareTo(bar)).isAtLeast(0)"
does not have any advantage compared to classical JUnit assertions.
If we use the richer assertion methods that Truth provides,
the code is easier to write, much easier to read,
and the error message will be much better because it will contain the
proper values of the interesting objects,
even without a custom message.

Furthermore, for this particular test we can even enhance it further
by not only testing for the lower bound but als for the upper bound.