disable proof production in Z3.
The option was required by the interpolation procedure.
This was completely removed from Z3, and thus the option is no longer required.

Additionally, this option avoids quantifier elimitation `qe-light`.
This might be a bug/regression in Z3 [1].
As long as we do not actively require proofs, we can disable them.

[1] https://github.com/Z3Prover/z3/commit/c9c9efebdeb48ee5ca11a5ca19b30c3db372c331#commitcomment-39911600