CVC4: fix deadlock for shutdownhook.
This commit simplifes the interrupt handling of CVC4TheoremProver.

First, we only interrupt long running native calls, e.g., the sat-check.
We do no longer interrupt push/pop/assert calls, which are quite fast anyway.

Second, there was a deadlock in the shutdown handling.
In detail: if the sat-check was fast and a shutdown was executed afterwards,
then the flag of the hook was never changed, causing an endless loop.
This looks like a deadlock without CPU usage,
because the loop executes Thread.sleep in each iteration.