update Princess to latest version.
The new Princess version comes from an official Maven repository.
This simplifies the installation process for our own Maven release.

The new Princess version contains some changes in array and quantifier support,
with several new types of formulas,
such as "IEquation" for Ints and "ExtArray" as new handler for Array theory.
We applied the changes to the corresponding components.