bugfix for the Princess bindings.
There is a bug in the Visitor/FormulaCreator for Princess.
For BV logic visiting a formula like `mod_cast(0, 4294967295, 2)`,
which models a value 2 with 32 bit via `mod_cast(lower, upper, value)`,
returns a function type `UF`. The correct function type would be `OTHER`.

The previous commit added the tests, this commit fixes the bug.

The bug was detected in https://gitlab.com/sosy-lab/software/cpachecker/-/issues/865 .
In CPAchecker's predicate analysis all symbols (variables and UFs)
are renamed when uninstantiating a formula after the interpolation.
The name does not need to change, but the renaming instantiates the wrong object,
because of the (wrong) function type UF.
The newly created UF instance is no longer related to the BV arithmetics and
allows arbitrary evaluations, causing a bug in the analysis.