Remove codacy integration
It seems to have stopped working some time ago,
the link behind the badges just gives 404.
The badges were still showing, but who knows whether their values were
outdated? For example, coverage shown was 60% but is actually 68%
according to our CI job.
We have removed Codacy from other projects as well because there was not
much value in it and it was a maintenance effort.

The badge for test coverage is replaced by the one that GitLab provides.
Note that the GitLab uses much worse colors (current badge is red).