Princess: change sort of bound variable in quanfified formula wiht no explicit bounds.
This commit fixes a bug appearing in CPAchecker:
Princess produces interpolants like `ALL !(3 * _0 = int_cast(y@3))` (representing `y@3 is not divisible by 3`).
When visiting such a formula in CPAchecker, we had changed the sort of the ound variable `_0` from `Int` to `Bool`.
The result was a repeated counterexample in predicate analysis, e.g., for a command line like:
```
scripts/cpa.sh -predicateAnalysis -setprop solver.solver=PRINCESS \
    -setprop cpa.predicate.encodeBitvectorAs=BITVECTOR -setprop cpa.predicate.encodeFloatAs=INTEGER \
    -spec test/programs/benchmarks/properties/unreach-call.prp -32 \
    test/programs/benchmarks/loop-acceleration/underapprox_1-2.c
```

I am not yet sure, whether Princess allows to use distinct or even multiple sorts for such quantified formulas.
For the user, having access to the bound variable in the visitor might be required. That is not yet possible here.