fix process to extract free variables from formulas: consider quantifier scopes.
Previously, we analysed all subformulas of the given formula in equal way and
ignored scopes from quantifiers, such that bound variables are also extracted.
With this commit, we start a new scope whenever visiting a quantified subformula
and take care of bounded variables, i.e. they are _not_ extracted from the formula.

Instead of a pure fixed-point algorithm for visiting all subformulas
(i.e., without any recursive call in the algorithm due to waitlist-usage),
we now have a recursive step if (and only if) we find a quantified subformula.
For other operations, we still use the non-recursive algorithm.