remove code for using a separate class-loader for Z3.
The code for the class-loader is quite old and might no longer be needed.
Z3 has a flag for disabling their own implementation for loading the native library,
and we set this flag before accessing any binary code of Z3.

This commit might remove one of the most complex snippets of code from JavaSMT.
All other solvers work without such a complex solution, as far as I know.

To be tested!