fix variable extraction in case of quantifiers.
The extraction process throws an exception,
if the same symbol is visited multiple times,
because the ImmutableMap-builder does not allow this.

We can choose to either allow redundant visitation,
or we avoid redundant visitation by checking symbol names.
We take the second approach and track all symbol names.

A JUnit test was added that triggers the exception.