Princess: rewrite method for dumping SMTLIB2.
The ordering of nested abbreviations was not considered for the export of SMTLIB2.
This can cause a crash when parsing such dumps, if nested abbreviations exist.
Now, we compute the correct ordering, such that
nested dependencies are dumped before the depending abbreviations.

Removing dependency from "abbrev_" as name of abbreviations.
Additionally, the ordering of normal symbols and UFs is deterministic now.