CVC4: disable some JUnit tests for theory combination of Arrays and Quantifiers.
There seems to be a bug in CVC4 that weakens its reasoning,
such that UNKNOWN/INCOMPLETE is returned from the SAT check.
As CVC4 is no longer actively maintained (there is already CVC5),
disabling the tests and having a basic support for String theory
is a quick and reasonable step.