refactoring(Boolector): remove ReusableStack-wrapper from Boolector.
Boolector is currently the only solver that effectively allows only one stack,
so we need to cleanup and reuse the stack.

There is no reason to wrap it in ReusableStackProver,
but we can directly implement the required operation in the constructor and on closing of a prover instance.
This simplifies the understanding and allows for easier debugging.