no longer show hint for other solver, if libraries are not available.
In different project setups, such projects build via Maven,
not only native libraries, but also Java libraries might not be available.
Thus, the message is misleading.
It is better to just show a general error message and
let the application itself provide hints hhow to solve the issue.

This change is related to #102.
At time of creating #102, the solvers MathSAT and Z3 were not available for non-Linux systems.
Currently, we provide several solvers, native libraries as well as Java-based ones.