fix exception for AND/OR-operations without arguments.
There are cases, where a solver (such as Z3) returns an empty AND/OR-operation,
e.g., as result of quantifier elimination.
The visitor crashed when accessing the non-existant first argument.

This fixes an exception in CPAchecker's JUnit-test
'ConfigurationFileChecks:instantiate_and_run[config/unmaintained/policy-bam.properties]'
and was noticed in https://gitlab.com/sosy-lab/software/cpachecker/-/issues/1007 .