update Z3 to version 3.12.1
The pre-compiled Linux version of Z3 3.12.1 requires glibc in version 2.35,
which is not available on older systems like Ubuntu 18.04, Ubuntu 20.04, Debian 10, or Debian 11.

As a solution, we provide a self-compiled Linux version depending on the older GLIBC 2.27.
This allows JavaSMT with Z3 to be executed on some more systems.

With the self-compiled library we can simplify the loading condition for Z3 in our JUnit tests.