BV theory: add new methods for BV rotation.
The simple case of rotation by constant integer is implemented for all solvers that support BV theory.
The complex case of rotation by another formula is not supported by all solvers,
but only by Z3 and Boolector, and thus only implemented or those solvers.

This is an imcompatible API change that provides several new methods for BV theory.