add script for building a Docker image for Ubuntu 22.04.
We want to avoid complex and time-consuming steps when building and maintaining Docker images,
such as compiling cmake from source,
whenever an SMT solver requires a dependency that is not natively available on a given Ubuntu version.
The number of JavaSMT users still using an Ubuntu version older than 22.04 might be quite low,
so we plan to use Ubuntu 22.04 for building upcoming releases of solver binaries.