Remove notification config from AppVeyor config file
We configure this in the AppVeyor settings on the webpage
instead of in the file. This prevents crawlers from accessing the mail
address and allows us to change the config for the whole repo at the
same time, instead of having open branches still send notifications to
an outdated address.