Strings: Restore behavior of JavaSMT 5.0.0 and only allow (printable) US-ASCII characters in String constants. Add a new option "solver.useUnicodeStrings" to enable Unicode support and allow arbitrary Strings.
The behavior of makeString() was underspecified if the String contains Unicode characters. Only Z3 would (sometimes) allow it, but CVC4 and CVC5 would crash. Princess (which was added after 5.0.0) uses Unicode internally and won't complain, but it also doesn't support SMTLIB escape sequences.

We now default to throwing an exception when Unicode characters are used in a String constant, and all escaping and unescaping for Princess is handled internally.

In addition to that we also added a new option "solver.useUnicodeString" to enable full support for Unicode in String constants. Again, the escaping and unescaping for all solvers is handled internally. Note, that in this mode SMTLIB escape sequences are ignored and will not be replaced by their Unicode character. To still use escape sequences the methods escapeUnicodeForSmtlib and unescapeUnicodeForSmtlib for AbstractStringFormulaManager can be used.