Strings: Remove precondition in the 'char' overload for StringFormulaManager.range
The precondition checks if the lower bound is smaller or equal to the upper bound. However, in SMTLIB re.range is defined for all inputs and will return the empty language if the interval is empty. Since both versions of StringFormulaManager.range() should behave the same, we remove the precondition.