MathSAT: Add special treatment for integer modulo when the divisor is zero
The solver may return any value if the divisor is zero. However, "modulo" is still a function and modulo(a,0) and modula(b,0) must evaluate to the same value whenever 'a' and 'b' are the same,

MathSAT: Add special treatment for integer modulo when the divisor is zero

The solver may return any value if the divisor is zero. However, "modulo" is still a function and modulo(a,0) and modula(b,0) must evaluate to the same value whenever 'a' and 'b' are the same,

MathSAT: Add special treatment for integer modulo when the divisor is zero

The solver may return any value if the divisor is zero. However, "modulo" is still a function and modulo(a,0) and modula(b,0) must evaluate to the same value whenever 'a' and 'b' are the same.