Fix last merge
The broken test always failed for CVC4, but the exception used to be thrown later. With the new changes we immediately throw an IllegalArgument exception when str.range is called with a bound that is not a single ASCII character.

This commit will fix the test to catch the new exception.