SmtInterpol: Fix function visitor
SMTInterpol allows indexed functions such as `extract` that take parameters in addition to their regular arguments. Since JavaSMT does not support such functions the parameters can't be accessed or changed by the function visitor. We simply consider them part of the function symbol and make sure that they are passed through when formulas are disassembled and recreated by the function visitor.