Visitor/Quantifier: mark visiting bound variables as deprecated.
We already replace all bound variables when visiting the body of a quantified formula,
so there should not be any bound variable left when visiting any formula,
and we can remove the code for this case.
Each formula or part of a formula that is visited is well-scoped,
such that all found variables are free variables.
The user needs to track scopes for quantified formulas on his own,
e.g., handle a stack of bound variables when visiting quantified formulas.