CVC4 Model: backport latest changes for CVC5 model-handling (including arrays and UFs) towards CVC4.
This change replaces the recursive implementation with faster and simpler visitor usage.
Additionally, we get proper array-handling in the CVC4 model.

This solution removes the ability to list UF assignments for bound variables as part of the model.
However, the model API is even documented in such a way that such values could be missing,
and the user should directly evaluate formulas in the model.