Model usage: fix usage of model values and model evaluation after changing the underlying context.
Several solvers referenced a live prover context in the model,
allowing the user to insert further queries in the model or remove queries from the context,
potentially making the model inconsistent and unreliable for further iteration or evaluation.

The solution consists of several steps, for some solvers:
- compute the value-assignments upfront when constructing the model.
- invalidate the model once the context is changed.

Solvers that already provide a persistent model, such as MathSAT, SMTInterpol, and Z3,
will not change their behaviour.
However, other solvers will throw an exception on invalid usage.

While this change is backward-compatible in the API,
it is a serious bugfix that changes the behaviour of some solver bindings in JavaSMT,
because we need to loosen a basic guarantee for models.